cleanup treemap component. create tree model module and types
